+++ extensions/raycast-wallpaper/CHANGELOG.md
@@ -2,6 +2,6 @@
 
 ## [Added Raycast Wallpaper] - 2022-04-20
 - Release of Raycast Wallpaper version 1.0
-- **Raycast Wallpaper**: Get and set Raycast wallpapers to desktop.
-- **Random Wallpaper**: Set random Raycast wallpapers to desktop.
+- **Set Raycast Wallpaper**: Get and set Raycast wallpapers to desktop.
+- **Roll Raycast Wallpaper**: Set random Raycast wallpapers to desktop.
 - **Capture Raycast Metadata**: Capture a screenshot for Raycast Extension's metadata.